+++ server/src/main/java/org/opensearch/action/admin/cluster/node/stats/NodeStats.java
@@ -239,6 +239,7 @@ public class NodeStats extends BaseNodeResponse implements ToXContentFragment {
             admissionControlStats = null;
         }
 
+        // TODO: Update version later
         if (in.getVersion().onOrAfter(Version.CURRENT)) {
             peerRecoveryStats = in.readOptionalWriteable(PeerRecoveryStats::new);
         } else {